==> notebooks/2. transform_original_data.ipynb <==
# Note
This Google Colab notebook is run in different sessions so the notebook is not suitable for running from start to finish in one go. This notebook serves more to "audit" my exploration.

%%capture
!pip install huggingface-hub polars==1.9.0 -U

%%capture
!huggingface-cli download HangenYuu/Steam_Games_Review --repo-type dataset --local-dir ./data

# Objectives
1. Normalize the data to make it look more like coming from an operational database.
2. Map the data to the correct schema to speed up initialization of the PosgreSQL database later.

# Transformation logic

1. Create Game ID in Games table.
2. Create tables for nested column (genres, developer, publisher).
3. Create bridge table for the nested column.
2. Create Users table with User ID & Username.
1. Use Game ID and User ID in Reviews table.



# Unnest `Games` table
**Note**: Exclude columns for system requirements as it's likely not be used in a SQL database (a document database such as MongoDB is suitable for the format). So they are dropped from the beginning.

import polars as pl
from pathlib import Path
import re

local_dir = Path("/content/data")

def parse_reviews(value):
    if "%" in value:
        # Extract percentage and total number
        match = re.search(r"(\d+)% of ([\d,]+)", value)
        if match:
            percentage = int(match.group(1))
            total = int(match.group(2).replace(",", ""))
            return int((percentage / 100) * total)
    else:
        # Extract the number directly
        match = re.search(r"\(([\d,]+)\)", value)
        if match:
            return int(match.group(1).replace(",", ""))

df = pl.scan_csv(local_dir / "games_description.csv").drop("minimum_system_requirement", "recommend_system_requirement", "link", "short_description")
df = df.with_row_index("game_id", 1)
df = df.with_columns(
    pl.col("genres").str.replace_many(["]", "'", "["], "").str.split(", "),
    pl.col("long_description").str.strip_prefix("About This Game").str.strip_chars(),
    pl.col("number_of_english_reviews").str.replace_all(",", "").cast(pl.Int32),
    pl.col(["developer", "publisher"])
    .str.replace_many(["]", "'", "["], "")
    .str.split(", "),
    pl.when(pl.col("release_date").str.contains(r"\d{1,2} \w{3}, \d{4}"))
    .then(pl.col("release_date").str.to_date("%d %b, %Y", strict=False))
    .otherwise(pl.col("release_date").str.to_date("%b %Y", strict=False))
    .alias("release_date"),
    pl.col("number_of_reviews_from_purchased_people").map_elements(
        parse_reviews, return_dtype=pl.Int32
    ),
)

df = df.with_columns(
    pl.when(pl.col("number_of_reviews_from_purchased_people").is_null())
    .then(pl.lit("Not enough data"))
    .otherwise(pl.col("overall_player_rating"))
    .cast(pl.Categorical("lexical"))
    .alias("overall_player_rating"),
    pl.when(pl.col("number_of_reviews_from_purchased_people").is_null())
    .then(pl.col("overall_player_rating").str.extract(r"(\d+)").cast(pl.Int32))
    .otherwise(pl.col("number_of_reviews_from_purchased_people"))
    .alias("number_of_reviews_from_purchased_people"),
)

df.collect()

df.collect_schema().names()

def create_dimension_n_mapping_tables(df: pl.LazyFrame, id_name: str, nested_column_name: str) -> tuple[pl.DataFrame, pl.DataFrame]:
    """
    Creates a game ID column in the original dataframe and generates genre dimension tables.

    Parameters:
    df: LazyFrame with a 'genres' column of type list[str]
    id_name: str, the name of the ID column in the new dimension table, without the "_id" suffix
    nested_column_name: str, the name of the nested column in the original dataframe

    Returns:
    tuple: (games_df, dimension_df, mapping_df)
        - dimension_df: DataFrame with genre_id and genre_name
        - mapping_df: DataFrame linking game_ids to genre_ids
    """
    exploded = (
        df.explode(nested_column_name)
    )

    dimension_df = (
        exploded
        .select(nested_column_name)
        .unique()
        .sort(nested_column_name)
        .with_row_index(f"{id_name}_id", 1)
        .rename({nested_column_name: id_name})
        .collect()
    )

    mapping_df = (
        exploded
        .join(
            dimension_df.lazy(),
            left_on=nested_column_name,
            right_on=id_name
        )
        .select(["game_id",f"{id_name}_id"])
        .collect()
    )

    return dimension_df, mapping_df

genre_dimension_df, genre_mapping_df = create_dimension_n_mapping_tables(df, "genre", "genres")

game_id = 1

original_genres = (
    df
    .filter(pl.col("game_id") == game_id)
    .select("genres")
    .collect()
    .item()
    .sort()
    .to_list()
)

mapped_genres = (
    genre_mapping_df
    .lazy()
    .filter(pl.col("game_id") == game_id)
    .join(
        genre_dimension_df.lazy(),
        left_on="genre_id",
        right_on="genre_id"
    )
    .select("genre")
    .collect()
    .get_column("genre")
    .sort()
    .to_list()
)

print("Original genres:", original_genres)
# e.g. ["FPS", "Moddable", "Shooter"]

print("Mapped genres:", mapped_genres)
# e.g. ["FPS", "Moddable", "Shooter"]

print("Matches:", set(original_genres) == set(mapped_genres))

developer_dimension_df, developer_mapping_df = create_dimension_n_mapping_tables(df, "developer", "developer")
publisher_dimension_df, publisher_mapping_df = create_dimension_n_mapping_tables(df, "publisher", "publisher")

df = df.drop("genres", "developer", "publisher")
df.collect()

## Final functions

def create_dimension_n_mapping_tables(df: pl.LazyFrame, id_name: str, nested_column_name: str) -> tuple[pl.LazyFrame, pl.LazyFrame]:
    """
    Creates a game ID column in the original dataframe and generates genre dimension tables.

    Parameters:
    df: LazyFrame with a 'genres' column of type list[str]
    id_name: str, the name of the ID column in the new dimension table, without the "_id" suffix
    nested_column_name: str, the name of the nested column in the original dataframe

    Returns:
    tuple: (games_df, dimension_df, mapping_df)
        - dimension_df: LazyFrame with id_name_id and id_name
        - mapping_df: LazyFrame linking game_ids to id_name_ids
    """
    exploded = (
        df.explode(nested_column_name)
    )

    dimension_df = (
        exploded
        .select(nested_column_name)
        .unique()
        .sort(nested_column_name)
        .with_row_index(f"{id_name}_id", 1)
        .rename({nested_column_name: id_name})
    )

    mapping_df = (
        exploded
        .join(
            dimension_df.lazy(),
            left_on=nested_column_name,
            right_on=id_name
        )
        .select(["game_id",f"{id_name}_id"])
    )

    return dimension_df, mapping_df

def process_games_description() -> None:
    def parse_reviews(value):
        if "%" in value:
            # Extract percentage and total number
            match = re.search(r"(\d+)% of ([\d,]+)", value)
            if match:
                percentage = int(match.group(1))
                total = int(match.group(2).replace(",", ""))
                return int((percentage / 100) * total)
        else:
            # Extract the number directly
            match = re.search(r"\(([\d,]+)\)", value)
            if match:
                return int(match.group(1).replace(",", ""))

    df = pl.scan_csv(local_dir / "games_description.csv").drop("minimum_system_requirement", "recommend_system_requirement", "link", "short_description")
    df = df.with_row_index("game_id", 1)
    df = df.with_columns(
        pl.col("genres").str.replace_many(["]", "'", "["], "").str.split(", "),
        pl.col("long_description").str.strip_prefix("About This Game").str.strip_chars(),
        pl.col("number_of_english_reviews").str.replace_all(",", "").cast(pl.Int32),
        pl.col(["developer", "publisher"])
        .str.replace_many(["]", "'", "["], "")
        .str.split(", "),
        pl.when(pl.col("release_date").str.contains(r"\d{1,2} \w{3}, \d{4}"))
        .then(pl.col("release_date").str.to_date("%d %b, %Y", strict=False))
        .otherwise(pl.col("release_date").str.to_date("%b %Y", strict=False))
        .alias("release_date"),
        pl.col("number_of_reviews_from_purchased_people").map_elements(
            parse_reviews, return_dtype=pl.Int32
        ),
    )

    df = df.with_columns(
        pl.when(pl.col("number_of_reviews_from_purchased_people").is_null())
        .then(pl.lit("Not enough data"))
        .otherwise(pl.col("overall_player_rating"))
        .cast(pl.Categorical("lexical"))
        .alias("overall_player_rating"),
        pl.when(pl.col("number_of_reviews_from_purchased_people").is_null())
        .then(pl.col("overall_player_rating").str.extract(r"(\d+)").cast(pl.Int32))
        .otherwise(pl.col("number_of_reviews_from_purchased_people"))
        .alias("number_of_reviews_from_purchased_people"),
    )
    genre_dimension_df, genre_mapping_df = create_dimension_n_mapping_tables(df, "genre", "genres")
    developer_dimension_df, developer_mapping_df = create_dimension_n_mapping_tables(df, "developer", "developer")
    publisher_dimension_df, publisher_mapping_df = create_dimension_n_mapping_tables(df, "publisher", "publisher")
    df = df.drop("genres", "developer", "publisher")
    destination_dir = local_dir / "processed"
    destination_dir.mkdir(exist_ok=True, parents=True)
    df.collect().write_parquet(destination_dir / "games_description.parquet")
    genre_dimension_df.collect().write_parquet(destination_dir / "genre_dimension.parquet")
    genre_mapping_df.collect().write_parquet(destination_dir / "genre_mapping.parquet")
    developer_dimension_df.collect().write_parquet(destination_dir / "developer_dimension.parquet")
    developer_mapping_df.collect().write_parquet(destination_dir / "developer_mapping.parquet")
    publisher_dimension_df.collect().write_parquet(destination_dir / "publisher_dimension.parquet")
    publisher_mapping_df.collect().write_parquet(destination_dir / "publisher_mapping.parquet")

process_games_description()

!huggingface-cli login

!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/raw ./raw --repo-type=dataset

!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/processed ./processed --repo-type=dataset

# Users Table from steam_game_reviews.csv
**Note**: The repo structure changed from the last operation, so the path changed.

%%capture
!pip install huggingface-hub[hf_transfer] polars==1.9.0 -U

%%capture
!huggingface-cli download HangenYuu/Steam_Games_Review --repo-type dataset --local-dir ./data

import polars as pl
from pathlib import Path
from datetime import datetime
import re

raw_dir = Path("/content/data/raw")
processed_dir = Path("/content/data/processed")

df = pl.scan_csv(raw_dir / "steam_game_reviews.csv", infer_schema_length=10000)

df = df.with_columns(
    pl.col("hours_played").str.replace_all(",", "").cast(pl.Float32),
    pl.col(["helpful", "funny"]).str.replace_all(",", "").cast(pl.Int64),
    pl.when(pl.col("recommendation") == "Recommended")
    .then(True)
    .otherwise(False)
    .alias("recommendation"),
    pl.when(
        ~(
            (pl.col("date").str.contains(r"\w+ \d{1,2},\s\d{4}"))
            | pl.col("date").str.contains(r"\d{1,2} \w+,\s\d{4}")
        )
    )
    .then(pl.concat_str([pl.col("date"), pl.lit(", " + str(datetime.now().year))]))
    .otherwise(pl.col("date"))
    .alias("date"),
    pl.when(pl.col("username").str.contains("\n"))
    .then(pl.col("username").str.extract(r"^(.*?)\n"))
    .otherwise(pl.col("username").fill_null("anonymous")),
)
df = df.with_columns(
    pl.when(pl.col("date").str.contains(r"\w+\s\d{1,2},\s\d{4}"))
    .then(pl.col("date").str.to_date("%B %d, %Y", strict=False))
    .when(pl.col("date").str.contains(r"\d{1,2}\s\w+,\s\d{4}"))
    .then(pl.col("date").str.to_date("%d %B, %Y", strict=False))
)

df.collect()

games_description_df = pl.scan_parquet(processed_dir / "games_description.parquet").select("game_id", "name")
games_description_df.collect()

users_df = (
    df
    .select("username")
    .unique()
    .with_row_index(f"user_id", 1)
)
users_df.collect()

updated_reviews_df = (
    df
    # Join with users to get user_id
    .join(
        users_df,
        left_on="username",
        right_on="username"
    )
    # Join with games to get game_id
    .join(
        games_description_df,
        left_on="game_name",
        right_on="name"
    )
    # Drop columns replaced with IDs
    .select(pl.all().exclude("game_name", "username"))
)
updated_reviews_df.collect()

## Final functions

import polars as pl
from pathlib import Path
from datetime import datetime
import re


def process_games_reviews() -> None:
    raw_dir = Path("/content/data/raw")
    processed_dir = Path("/content/data/processed")

    df = pl.scan_csv(raw_dir / "steam_game_reviews.csv", infer_schema_length=10000)
    df = df.with_columns(
        pl.col("hours_played").str.replace_all(",", "").cast(pl.Float32),
        pl.col(["helpful", "funny"]).str.replace_all(",", "").cast(pl.Int64),
        pl.when(pl.col("recommendation") == "Recommended")
        .then(True)
        .otherwise(False)
        .alias("recommendation"),
        pl.when(
            ~(
                (pl.col("date").str.contains(r"\w+ \d{1,2},\s\d{4}"))
                | pl.col("date").str.contains(r"\d{1,2} \w+,\s\d{4}")
            )
        )
        .then(pl.concat_str([pl.col("date"), pl.lit(", " + str(datetime.now().year))]))
        .otherwise(pl.col("date"))
        .alias("date"),
        pl.when(pl.col("username").str.contains("\n"))
        .then(pl.col("username").str.extract(r"^(.*?)\n"))
        .otherwise(pl.col("username").fill_null("anonymous")),
    )
    df = df.with_columns(
        pl.when(pl.col("date").str.contains(r"\w+\s\d{1,2},\s\d{4}"))
        .then(pl.col("date").str.to_date("%B %d, %Y", strict=False))
        .when(pl.col("date").str.contains(r"\d{1,2}\s\w+,\s\d{4}"))
        .then(pl.col("date").str.to_date("%d %B, %Y", strict=False))
    )

    games_description_df = pl.scan_parquet(
        processed_dir / "games_description.parquet"
    ).select("game_id", "name")

    users_df = df.select("username").unique().with_row_index(f"user_id", 1)

    updated_reviews_df = (
        df
        # Join with users to get user_id
        .join(users_df, left_on="username", right_on="username")
        # Join with games to get game_id
        .join(games_description_df, left_on="game_name", right_on="name")
        # Drop columns replaced with IDs
        .select(pl.all().exclude("game_name", "username"))
    )
    updated_reviews_df.collect().write_parquet(processed_dir / "games_reviews.parquet")
    users_df.collect().write_parquet(processed_dir / "users_mapping.parquet")

process_games_reviews()

!export HF_HUB_ENABLE_HF_TRANSFER=1
!huggingface-cli login
!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/raw ./raw --repo-type=dataset
!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/processed ./processed --repo-type=dataset

# Ranking table
This table is very simple.

import polars as pl
from pathlib import Path


def process_games_ranking() -> None:
    raw_dir = Path("/content/data/raw")
    processed_dir = Path("/content/data/processed")

    genre_dimension_df = pl.scan_parquet(processed_dir / "genre_dimension.parquet")
    games_description_df = pl.scan_parquet(processed_dir / "games_description.parquet").select("game_id", "name")

    schema = pl.Schema(
        {
            "game_name": pl.String(),
            "genre": pl.String(),
            "rank_type": pl.String(),
            "rank": pl.UInt8(),
        }
    )

    df = pl.scan_csv(raw_dir / "games_ranking.csv", schema=schema)
    rank_df = df.select("rank_type").unique().with_row_index("rank_type_id", 1)

    df = (
        df.join(
            games_description_df,
            left_on="game_name",
            right_on="name",
        )
        .join(genre_dimension_df, left_on="genre", right_on="genre")
        .join(rank_df, left_on="rank_type", right_on="rank_type")
    )

    df.select(
        pl.all().exclude("game_name", "genre", "rank_type")
    ).collect().write_parquet(processed_dir / "games_ranking.parquet")
    rank_df.collect().write_parquet(processed_dir / "rank_dimension.parquet")

process_games_ranking()

pl.read_parquet(processed_dir / "games_ranking.parquet")

!export HF_HUB_ENABLE_HF_TRANSFER=1
!huggingface-cli login
!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/raw ./raw --repo-type=dataset
!huggingface-cli upload HangenYuu/Steam_Games_Review /content/data/processed ./processed --repo-type=dataset
