==> notebooks/3. postgres_tables_check.ipynb <==
import polars as pl
from sqlalchemy import create_engine

engine = create_engine("postgresql://postgres:postgres@localhost:5432/postgres")

engine.connect()

query = """
SELECT *
FROM pg_catalog.pg_tables
"""

pl.read_database(query, connection=engine)

query = """
SELECT *
FROM pg_catalog.pg_tables
WHERE schemaname != 'pg_catalog' AND 
    schemaname != 'information_schema';
"""

tables = list(
    pl.read_database(query, connection=engine)
    .select("tablename")
    .to_pandas()
    .values.flatten()
)

tables

for table in tables:
    query = f"""
    SELECT * FROM recsys.{table} LIMIT 10
    """

    df = pl.read_database(query, connection=engine)
    print(df.schema)

# All tables ingested successfully and schemas are correct (except for date columns, which should be handle in)!
