==> notebooks/4. collaborative_filtering_example.ipynb <==
# Collaborative Filtering (ALS)

!pip install huggingface_hub[hf_transfer] -U
!export HF_HUB_ENABLE_HF_TRANSFER=1
!huggingface-cli download HangenYuu/Steam_Games_Review processed/games_reviews.parquet --repo-type dataset --local-dir .

from pyspark.sql import SparkSession
from pyspark.ml.recommendation import ALS
from pyspark.sql import functions as F

spark = SparkSession.builder.appName("RecommendationSystem").getOrCreate()

# Load your data
data = spark.read.parquet("/content/processed/games_reviews.parquet", header=True, inferSchema=True)

data = data.select('user_id', 'game_id', 'recommendation')
data = data.withColumn('recommendation_numeric', F.col('recommendation').cast('int')).drop('recommendation')

# Create an ALS model
als = ALS(
    maxIter=10,
    regParam=0.1,
    userCol='user_id',
    itemCol='game_id',
    ratingCol='recommendation_numeric',
    coldStartStrategy='drop'
)

model = als.fit(data)

user_recommendations = model.recommendForAllUsers(10)
user_recommendations.coalesce(1).write.parquet("/content/user_recommendations.parquet", mode="overwrite")

!huggingface-cli login

!export HF_HUB_ENABLE_HF_TRANSFER=1
!huggingface-cli upload HangenYuu/Steam_Games_Review /content/user_recommendations.parquet/part-00000-674a66e6-9a1e-4855-b492-c2557d6577f1-c000.snappy.parquet ./results/user_recommendations.parquet --repo-type=dataset
