==> notebooks/5. results_data_transformation.ipynb <==
import pandas as pd

df = pd.read_parquet("../data_pipeline/data/results/user_recommendations.parquet")
df.head()

df["recommendations"] = df["recommendations"].apply(
    lambda x: [item["game_id"] for item in x]
)
df.head()

records = df.to_dict(orient="records")

records[96]
